requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.9.0
